+++ Source/3rdParty/mupen64plus-core/src/device/rdram/rdram.h
@@ -52,7 +52,7 @@ struct rdram
     uint32_t regs[RDRAM_MAX_MODULES_COUNT][RDRAM_REGS_COUNT];
 
     uint32_t* dram;
-    size_t dram_size;
+    uint32_t dram_size;
 
     uint8_t corrupted_handler;
 
@@ -71,7 +71,7 @@ static osal_inline uint32_t rdram_dram_address(uint32_t address)
 
 void init_rdram(struct rdram* rdram,
                 uint32_t* dram,
-                size_t dram_size,
+                uint32_t dram_size,
                 struct r4300_core* r4300);
 
 void poweron_rdram(struct rdram* rdram);
